+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Het sluitknopje is verplaatst, zie PR voor geldige markup ([#1765](https://githu
 `Progress Bar` en `Progress Indicator` labels staan nu als eerste element en worden gevolgd door het visuele `progress` kenmerk ([#1748](https://github.com/dso-toolkit/dso-toolkit/pull/1748))
 * **core + css + sources:** /sources variables en mixins verhuizen naar /sources/components ([#1777](https://github.com/dso-toolkit/dso-toolkit/issues/1777))
 * **css + dso-toolkit:** Gedateerde componenten/styling opruimen ([#1458](https://github.com/dso-toolkit/dso-toolkit/issues/1458))
+* **core + sources + storybook:** Autosuggest: null kunnen meegeven aan suggestions voor 'nog niet gezocht' ([#1784](https://github.com/dso-toolkit/dso-toolkit/issues/1784))
 
 ## 44.0.0